fix: very obscure bug!!!
the issue is that blink.cmp's `<cr>` to accept completion keybind stops
triggering after running `:e` on a markdown file. the specific reason is
because of an interplay between `autoclose.nvim`, `blink.cmp` and my
autocmds which sets maps `<cr>` to a `bullets-nvim` command on
`BufReadPost`.

I am not completely sure exactly what the cause is in that whole
interplay, but I am certain all of those plugins are involved as if I
disable `autoclose.nvim`, the issue goes away, which is somehow not the
case when I remove the keybind to `<cr>` explicitly.

therefore, a simple workaround is to prevent the markdown keybinds from
rebinding every time the buffer is reloaded but to only bind when the
buffer is loaded initially. This is technically the behavior which
should have been occuring at the start so it is not technically a
workaround but a bug fix.